+++ packages/block/src/header.ts
@@ -459,7 +459,8 @@ export class BlockHeader {
     let parentGasLimit = parentBlockHeader.gasLimit
     // EIP-1559: assume double the parent gas limit on fork block
     // to adopt to the new gas target centered logic
-    if (this.number.eq(this._common.hardforkBlockBN('london'))) {
+    const londonHardforkBlock = this._common.hardforkBlockBN('london')
+    if (londonHardforkBlock && this.number.eq(londonHardforkBlock)) {
       const elasticity = new BN(this._common.param('gasConfig', 'elasticityMultiplier'))
       parentGasLimit = parentGasLimit.mul(elasticity)
     }
@@ -592,7 +593,8 @@ export class BlockHeader {
       if (!this.baseFeePerGas) {
         throw new Error('EIP1559 block has no base fee field')
       }
-      const isInitialEIP1559Block = this.number.eq(this._common.hardforkBlockBN('london'))
+      const block = this._common.hardforkBlockBN('london')
+      const isInitialEIP1559Block = block && this.number.eq(block)
       if (isInitialEIP1559Block) {
         const initialBaseFee = new BN(this._common.param('gasConfig', 'initialBaseFee'))
         if (!this.baseFeePerGas!.eq(initialBaseFee)) {
@@ -885,7 +887,7 @@ export class BlockHeader {
     if (this._common.hardforkIsActiveOnChain('dao')) {
       // verify the extraData field.
       const blockNumber = this.number
-      const DAOActivationBlock = this._common.hardforkBlockBN('dao')
+      const DAOActivationBlock = this._common.hardforkBlockBN('dao')!
       if (blockNumber.gte(DAOActivationBlock)) {
         const drift = blockNumber.sub(DAOActivationBlock)
         if (drift.lte(DAO_ForceExtraDataRange)) {