+++ packages/common/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 (modification: no type change headlines) and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+**New Features**
+
+**Bug Fixes and Maintenance**
+
+- Fixed `hardforkBlockBN()` to correctly return null for unscheduled hardforks, PR [#1329](https://github.com/ethereumjs/ethereumjs-monorepo/pull/1329)
+
+**Dependencies, CI and Docs**
+
 ## 2.4.0 - 2021-07-08
 
 ### Finalized London HF Support